requirements.txt
streamlit==1.31.0
requests==2.32.3
psycopg2-binary==2.9.9
numpy==1.26.4
